requirements.txt
Django==3.0.2
feedparser==5.2.1
mf2py==1.1.2
bleach==3.1.3
gunicorn==20.0.4
beautifulsoup4==4.9.0
selenium==3.141.0
python-dateutil==2.8.1